Renaming components/basic/misc to components/basic/display
