+++ src/components/basic/display/TokenDisplay/TokenDisplay.stories.tsx
@@ -20,7 +20,7 @@ const mockContext = {
 
 export default {
   component: TokenDisplay,
-  title: "basic/misc/TokenDisplay",
+  title: "basic/display/TokenDisplay",
   // Our exports that end in "Data" are not stories.
   excludeStories: /.*Data$/,
   decorators: [
