+++ src/components/basic/inputs/FundingInput/PerBracketAmount.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-import { TokenDisplay } from "components/basic/misc/TokenDisplay";
-import { SubtextAmount } from "components/basic/misc/SubtextAmount";
+import { TokenDisplay } from "components/basic/display/TokenDisplay";
+import { SubtextAmount } from "components/basic/display/SubtextAmount";
 
 export interface Props {
   amount: string;
